Refactor testing app to testutil
Ref oursky/likecoin-chain#124
